// src/app/models/app.ts
export default interface App {
    id: number;
    rank:number;
    class:string;
    arrow:string;
    change:number;
    title: string;
    icon: string;
    rating: 0;
    reviews: 0;
    newprice: string;
    oldprice: string;
    link: string;
    genres: any;
    dev: string;
    isActive: boolean;
    visible: boolean;
    state:string;
    catNames:string;
    maxWidth:string;
    device: number;
  }
  